+++ NEWS.md
@@ -1,6 +1,7 @@
 # httr2 (development version)
 
 * Export `is_online()` as thin wrapper around `curl::has_internet()` (#512).
+* `curl_translate()` now translates cookie headers to `req_cookies_set()` (#431).
 * `curl_transform()` will now use `req_body_json_modify()` for JSON data (#258).
 * `resp_stream_is_complete()` tells you if there is still data remaining to be streamed (#559).
 * New `url_modify()`, `url_modify_query()`, and `url_modify_relative()` make it easier to modify an existing url (#464).